# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API de Gestión de Espacios de Coworking",
    "description": "Documentación de la API para Autenticación, Gestión de Espacios, Reservas y Reportes.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/reservation": {
      "post": {
        "tags": [
          "Reservations"
        ],
        "summary": "Create a new reservation",
        "description": "Creates a reservation executing validations and payment processing",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reserva creada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralResponse"
                }
              }
            }
          },
          "400": {
            "description": "Pago rechazado o duración inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralResponseWithErrors"
                }
              }
            }
          },
          "409": {
            "description": "El espacio ya está reservado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralResponseWithErrors"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Reservations"
        ],
        "summary": "Get current user reservations",
        "description": "Returns all reservations belonging to the authenticated user.",
        "responses": {
          "200": {
            "description": "Lista de reservas del usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/reservation/{reservationId}/cancel": {
      "patch": {
        "tags": [
          "Reservations"
        ],
        "summary": "Cancel reservation",
        "description": "Cancels a reservation owned by the authenticated user. ADMIN can cancel any reservation.",
        "parameters": [
          {
            "name": "reservationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reserva cancelada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralResponse"
                }
              }
            }
          },
          "404": {
            "description": "Reserva no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralResponseWithErrors"
                }
              }
            }
          },
          "409": {
            "description": "La reserva no puede ser cancelada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralResponseWithErrors"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PaymentMethod": {
        "type": "string",
        "enum": [
          "CREDIT_CARD",
          "DEBIT_CARD",
          "BANK_TRANSFER"
        ]
      },
...
      "GeneralResponse": {
        "type": "object",
        "description": "Respuesta estándar de la API.",
        "properties": {
          "uri": {
            "type": "string",
            "description": "URI del recurso solicitado.",
            "example": "http://localhost:8080/api/v1/balance/process-transaction"
          },
          "message": {
            "type": "string",
            "description": "Mensaje descriptivo de la operación.",
            "example": "Transacción recibida correctamente."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "Código HTTP de la respuesta.",
            "example": 202
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha y hora de generación de la respuesta."
          },
          "data": {
            "type": "object",
            "description": "Información adicional de la respuesta."
          }
        }
      },
      "GeneralResponseWithErrors": {
        "allOf": [
          {
            "$ref": "#/components/schemas/GeneralResponse"
          },
          {
            "type": "object",
            "description": "Respuesta utilizada cuando ocurre uno o varios errores.",
            "properties": {
              "errors": {
                "type": "array",
                "description": "Lista de errores encontrados",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        ]
      },
...
      "CreateReservationRequest": {
        "type": "object",
        "required": [
          "spaceId",
          "startTime",
          "endTime",
          "paymentMethod"
        ],
        "properties": {
          "spaceId": {
            "type": "integer",
            "format": "int64"
          },
          "startTime": {
            "type": "string",
            "example": "2023-10-25 14:30",
            "description": "Fecha y hora de inicio (Formato yyyy-MM-dd HH:mm)"
          },
          "endTime": {
            "type": "string",
            "example": "2023-10-25 16:30",
            "description": "Fecha y hora de fin (Formato yyyy-MM-dd HH:mm)"
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/PaymentMethod"
          }
        }
      },
...
    }
  }
}