# src/main/java/com/company/coworking/management/dto/response/GeneralResponse.java
package com.company.coworking.management.dto.response;

import com.fasterxml.jackson.annotation.JsonInclude;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class GeneralResponse {

    private String uri;

    @JsonInclude(JsonInclude.Include.NON_NULL)
    private String message;

    private int status;

    private LocalDateTime timestamp;

    @JsonInclude(JsonInclude.Include.NON_NULL)
    private Object data;
}